Restructure file structure
Restructure the project file structure that is conventional pattern for library projects.
