+++ h264p/include/h264p/sei.h
@@ -4,7 +4,7 @@
 #include <vector>
 #include <memory>
 
-#include "loki/Visitor.h"
+#include <loki/Visitor.h>
 
 namespace ThetaStream
 {
