+++ src/MDPMParser.h
@@ -1,7 +1,7 @@
 #pragma once
-#include "loki/Visitor.h"
-#include "seiimpl.h"
-#include "json/elements.h"
+#include <loki/Visitor.h>
+#include <h264p/seiimpl.h>
+#include <json/elements.h>
 
 #include <memory.h>
 #include <vector>
